So, i've spent the day rewriting the framework so that the Model is split up into an Object and Data Mapper. I've found Abstract in Java to be really difficult in comparison  to PHP, for example, it's apparently impossible to write a shared method that can for example retrieve all entries from a Table. So, all i've done is defined the methods that need abstracting in the GPSISDataMapper class.
For all Objects there will be 2 constructors. One that includes an id as a parameter, and one that creates a new object and calls the put method to insert the new object into the table as soon as it's created.

For the Database, to make our lives a lot easier, so we DON'T need to use port tunneling via SSH and we can configure the Database via phpMyAdmin (an exceptionally useful GUI for MySQL), i've created a Database on my VPS, for us all to use. You can access it here: https://zpan.alineofcode.co.uk/etc/apps/phpmyadmin/ with:
username: grpj
password: wv7PBUZBMXH88NP6

When we get towards the end of our implementation, we can simply export the whole database and import into the one that we've been provided at campus :)

So now, our folder structure looks like
exception/
....Any custom Exceptions we use will go in here, there'll be quite a few when we get to the interface
framework/
...GPSIS.java
...GPSISDataMapper.java
...GPSISFramework.java
...GPSISModule.java
...GPSISObject.java
mapper/
...CareProgrammeDMO.java
...RoomDMO.java
module/
...LoginModule.java
...MenuModule.java
object/
...CareProgramme.java
...Room.java
template/
....Still haven't taken a look at Java Swing fully, but as i've been saying for a while now, it'd be nice for us to template our Views so we keep the whole application visually consistent :)

I'll push these changes into the trunk now